wallet: Separate the `AvailableCoins()` function into a new file
This allows `AvailableCoins()` to be used in multiple files
without creating a circular dependency
